Fix: Redirect the homepage logic from home directory to dashboard root
